backend/requirements.txt
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.32.0,<1.0.0
gunicorn>=23.0.0,<24.0.0
pydantic[email]>=2.10.0,<3.0.0
pydantic-settings>=2.7.0,<3.0.0
supabase>=2.11.0,<3.0.0
openai>=1.60.0,<2.0.0
sse-starlette>=2.2.0,<3.0.0
slowapi>=0.1.9,<1.0.0
pyyaml>=6.0.0,<7.0.0
python-jose[cryptography]>=3.3.0,<4.0.0
bcrypt>=4.0.0,<5.0.0
